04看图猜成语微信小程序/guess_idiom/fastapi_idiom/requirements.txt
# bcrypt
# certifi
# cffi
# chardet
# click
# fastapi
# h11
# httptools
# idna
# passlib
# pycparser
# pydantic
# PyJWT
# PyMySQL
# python-multipart
# requests
# six
# SQLAlchemy
# starlette
# urllib3
# uvicorn
# websockets
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
idna==2.9
passlib==1.7.2
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
PyMySQL==0.9.3
python-multipart==0.0.5
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.3
websockets==8.1

